# api/src/openapi.yml
openapi: 3.0.3
info:
  title: Dawson POS API
  version: 1.0.0
  description: >
    API para sistema de ventas, stock, caja, clientes, proveedores, compras, combos dinámicos,
    con control estricto por roles (ADMIN / CAJERO), dinero en centavos (integer) y combos por coincidencia exacta.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /auth/me:
    get:
      tags: [Auth]
      summary: Usuario actual
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MeResponse" }
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /clients:
    get:
      tags: [Clients]
      summary: Buscar/listar clientes (búsqueda por nombre)
      parameters:
        - in: query
          name: search
          schema: { type: string }
          description: Búsqueda parcial por nombre
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedClients" }
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    post:
      tags: [Clients]
      summary: Crear cliente (modal rápido POS)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateClientRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Client" }
        "409":
          description: Teléfono duplicado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PhoneAlreadyExistsError" }
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /clients/{clientId}:
    get:
      tags: [Clients]
      summary: Obtener cliente
      parameters:
        - $ref: "#/components/parameters/ClientId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Client" }
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    patch:
      tags: [Clients]
      summary: Editar cliente
      parameters:
        - $ref: "#/components/parameters/ClientId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateClientRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Client" }
        "409":
          description: Teléfono duplicado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PhoneAlreadyExistsError" }
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /categories:
    get:
      tags: [Categories]
      summary: (ADMIN) Listar categorías
      parameters:
        - in: query
          name: includeInactive
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CategoryListResponse" }
        "403": { $ref: "#/components/responses/ForbiddenError" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
    post:
      tags: [Categories]
      summary: (ADMIN) Crear categoría
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCategoryRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "403": { $ref: "#/components/responses/ForbiddenError" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
  /categories/{categoryId}:
    patch:
      tags: [Categories]
      summary: (ADMIN) Editar categoría
      parameters:
        - in: path
          name: categoryId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCategoryRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "403": { $ref: "#/components/responses/ForbiddenError" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
    delete:
      tags: [Categories]
      summary: (ADMIN) Eliminar categoría (soft/hard)
      parameters:
        - in: path
          name: categoryId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeleteResult" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "403": { $ref: "#/components/responses/ForbiddenError" }
        "401": { $ref: "#/components/responses/UnauthorizedError" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      in: query
      name: pageSize
      schema: { type: integer, minimum: 1, maximum: 200, default: 25 }
    ClientId:
      in: path
      name: clientId
      required: true
      schema: { type: string }
...
  responses:
    UnauthorizedError:
      description: No autenticado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: unauthorized
            message: Token inválido o ausente
    ForbiddenError:
      description: Sin permisos (rol)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: forbidden
            message: No tiene permisos para esta acción
    NotFoundError:
      description: No encontrado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: not_found
            message: Recurso no encontrado
...
  schemas:
...
    Role:
      type: string
      enum: [ADMIN, CAJERO]
...
    LoginRequest:
      type: object
      required: [usuario, password]
      properties:
        usuario: { type: string }
        password: { type: string }
    LoginResponse:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user:
          type: object
          required: [id, nombre, rol]
          properties:
            id: { type: string }
            nombre: { type: string }
            rol: { $ref: "#/components/schemas/Role" }
    MeResponse:
      type: object
      required: [id, nombre, rol]
      properties:
        id: { type: string }
        nombre: { type: string }
        rol: { $ref: "#/components/schemas/Role" }
    Client:
      type: object
      required: [id, nombre, telefono, direccion]
      properties:
        id: { type: string }
        nombre: { type: string }
        telefono: { type: string }
        direccion: { type: string }
    CreateClientRequest:
      type: object
      required: [nombre, telefono, direccion]
      properties:
        nombre: { type: string, minLength: 1 }
        telefono: { type: string, minLength: 3 }
        direccion: { type: string, minLength: 1 }
    UpdateClientRequest:
      type: object
      properties:
        nombre: { type: string }
        telefono: { type: string }
        direccion: { type: string }
    PaginatedClients:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Client" }
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
    PhoneAlreadyExistsError:
      type: object
      required: [error, message, existingClientId]
      properties:
        error: { type: string, enum: [phone_already_exists] }
        message: { type: string }
        existingClientId: { type: string }
...
    Error:
      type: object
      required: [error, message]
      properties:
        error: { type: string }
        message: { type: string }
...